Allow information from backends to be used when instantiating the player.
For backends, and thus authentication mechanisms, to work you would need to:

1. authenticate in the client using a backend
2. send a player identifier and verification to a game logic worker
3. Verify the player identifier using the verification on the game logic worker
   with the game's backend to prevent impersonation attacks.
4. retrieve player information from the game's backend
5. use the retrieved information to create / populate the components of the
   player entity while creating it.

With the changes in this commit the game consuming the GDK is able to change
the create player command's request to contain data (using the `extra_data`
field) as they see fit and use a delegate named `GetPlayerCreationData` of
the `PlayerLifecycleConfig` to create a payload of type
Improbable.PlayerLifecycle.CreatePlayerRequestType.

This payload is passed onto the player creation command and is received by the
delegate `CreatePlayerEntityTemplate`. Then during the execution of this
delegate, which is controlled by the consumer, the consumer may query the
backend using the data in the payload.

At this point the consumer may reject the data and stop the connection, or
construct an entity with the components prepopulated with data from the
backend.